Siden alt av grunnlag går gjennom behndling kan vi fjerne grunnlagsklienten i tidshendelser
