+++ apps/etterlatte-tidshendelser/src/test/kotlin/no/nav/etterlatte/tidshendelser/OmstillingsstoenadDoedsdatoTest.kt
@@ -14,7 +14,6 @@ import no.nav.etterlatte.tidshendelser.hendelser.HendelserJobb
 import no.nav.etterlatte.tidshendelser.hendelser.JobbStatus
 import no.nav.etterlatte.tidshendelser.hendelser.Steg
 import no.nav.etterlatte.tidshendelser.klient.BehandlingKlient
-import no.nav.etterlatte.tidshendelser.klient.GrunnlagKlient
 import no.nav.etterlatte.tidshendelser.omstillingsstoenad.OmstillingsstoenadService
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
@@ -23,23 +22,22 @@ import java.time.Month
 import java.time.YearMonth
 
 class OmstillingsstoenadDoedsdatoTest {
-    private val grunnlagKlient: GrunnlagKlient = mockk<GrunnlagKlient>()
     private val behandlingKlient: BehandlingKlient = mockk<BehandlingKlient>()
     private val hendelseDao: HendelseDao = mockk<HendelseDao>()
-    private val omstillingsstoenadService = OmstillingsstoenadService(hendelseDao, grunnlagKlient, behandlingKlient)
+    private val omstillingsstoenadService = OmstillingsstoenadService(hendelseDao, behandlingKlient)
 
     @Test
     fun `ingen saker for aktuell maaned gir ingen hendelser`() {
         val behandlingsmaaned = YearMonth.of(2024, Month.APRIL)
         val jobb = hendelserJobb(JobbType.OMS_DOED_3AAR, behandlingsmaaned)
 
-        every { grunnlagKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) } returns emptyList()
+        every { behandlingKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) } returns emptyList()
         every { behandlingKlient.hentSaker(emptyList()) } returns emptyMap()
         every { behandlingKlient.hentSakerForPleieforholdetOpphoerte(any()) } returns emptyList()
 
         omstillingsstoenadService.execute(jobb)
 
-        verify { grunnlagKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) }
+        verify { behandlingKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) }
         verify(exactly = 0) { hendelseDao.opprettHendelserForSaker(jobb.id, any(), any()) }
     }
 
@@ -50,7 +48,7 @@ class OmstillingsstoenadDoedsdatoTest {
         val sakIder: List<SakId> = listOf(sakId1, sakId2, sakId3)
         val saker = (sakIder + SakId(4)).map { sak(it, SakType.OMSTILLINGSSTOENAD) }.associateBy { it.id }
 
-        every { grunnlagKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) } returns sakIder
+        every { behandlingKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) } returns sakIder
         every { behandlingKlient.hentSaker(sakIder + SakId(4L)) } returns saker
         every { behandlingKlient.hentSakerForPleieforholdetOpphoerte(any()) } returns listOf(SakId(4L))
         every { hendelseDao.opprettHendelserForSaker(jobb.id, listOf(sakId1, sakId2, sakId3, SakId(4L)), Steg.IDENTIFISERT_SAK) } returns
@@ -58,7 +56,7 @@ class OmstillingsstoenadDoedsdatoTest {
 
         omstillingsstoenadService.execute(jobb)
 
-        verify { grunnlagKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) }
+        verify { behandlingKlient.hentSakerForDoedsfall(behandlingsmaaned.minusYears(3)) }
         verify { hendelseDao.opprettHendelserForSaker(jobb.id, listOf(sakId1, sakId2, sakId3, SakId(4L)), Steg.IDENTIFISERT_SAK) }
     }
